services/ocr/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdf2image
Pillow
pydantic
python-dotenv
pypdf
